@@ -20,17 +20,22 @@
     "irc": "irc://irc.freenode.net/imbo"
   },
   "require": {
-    "php": ">=5.6",
-    "behat/behat": "^3.0",
-    "guzzlehttp/guzzle": "^6.0",
-    "beberlei/assert": "^2.1",
-    "firebase/php-jwt": "^4.0 | ^5.0"
+    "php": "^7.4 || ^8.0",
+    "ext-json": "*",
+    "beberlei/assert": "^3.3",
+    "behat/behat": "^3.8",
+    "firebase/php-jwt": "^5.2",
+    "guzzlehttp/guzzle": "^7.3"
   },
   "require-dev": {
-    "silex/silex": "^2.0",
-    "symfony/security": "^3.1",
-    "symfony/process": "^3.1",
-    "phpunit/phpunit": "^5.5"
+    "alexeyshockov/guzzle-psalm-plugin": "^0.3.1",
+    "phpunit/phpunit": "^9.5",
+    "psalm/plugin-phpunit": "^0.15.1",
+    "slim/psr7": "^1.3",
+    "slim/slim": "^4.7",
+    "symfony/process": "^5.2",
+    "tuupola/slim-basic-auth": "^3.3",
+    "vimeo/psalm": "^4.7"
   },
   "autoload": {
     "psr-4": {
@@ -43,16 +48,27 @@
     }
   },
   "scripts": {
-    "clean": "rm -rf build",
     "lint": "for file in `git ls-files '*php'`; do php -l $file; done",
-    "test-phpunit": "vendor/bin/phpunit",
-    "test-phpunit-coverage": "vendor/bin/phpunit --coverage-html build/coverage",
-    "test-behat": "vendor/bin/behat --strict",
+    "phpunit": "vendor/bin/phpunit",
+    "phpunit:coverage": "vendor/bin/phpunit --coverage-html build/coverage",
+    "behat": "vendor/bin/behat --strict",
+    "psalm": "vendor/bin/psalm",
     "test": [
-      "@test-phpunit",
-      "@test-behat"
+      "@phpunit",
+      "@behat"
+    ],
+    "sa": [
+      "@psalm"
+    ],
+    "ci": [
+      "@lint",
+      "@test",
+      "@sa"
     ],
     "dev": "php -S localhost:8080 -t ./features/bootstrap > server.log 2>&1",
     "docs": "cd docs; make html"
+  },
+  "config": {
+    "sort-packages": true
   }
 }
