@@ -1,36 +1,40 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\Context;
 
-use Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\Jwt as JwtMatcher;
+use Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\JWT as JwtMatcher;
 use Imbo\BehatApiExtension\ArrayContainsComparator;
 use Imbo\BehatApiExtension\Exception\AssertionFailedException;
-use Behat\Behat\Context\SnippetAcceptingContext;
 use Behat\Gherkin\Node\PyStringNode;
 use Behat\Gherkin\Node\TableNode;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Request;
-use GuzzleHttp\Psr7;
+use GuzzleHttp\Psr7\UriResolver;
 use Assert\Assertion;
 use Assert\AssertionFailedException as AssertionFailure;
+use GuzzleHttp\Psr7\Utils;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\UriInterface;
 use InvalidArgumentException;
 use RuntimeException;
 use stdClass;
 
 /**
  * Behat feature context that can be used to simplify testing of JSON-based RESTful HTTP APIs
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
-class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareContext, SnippetAcceptingContext {
+class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareContext {
     /**
      * Guzzle client
      *
      * @var ClientInterface
      */
-    protected $client;
+    protected ClientInterface $client;
+
+    /**
+     * Base URI used by the Guzzle client
+     */
+    protected string $baseUri;
 
     /**
      * Request instance
@@ -46,16 +50,26 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * Options to send with the request.
      *
-     * @var array
+     * @var array{
+     *   auth: string[],
+     *   form_params: array<string, string|string[]>,
+     *   multipart: array<int, array{name: string, contents: string|resource}>,
+     *   query: array<string, mixed>
+     * }
      */
-    protected $requestOptions = [];
+    protected $requestOptions = [
+        'auth'        => [],
+        'form_params' => [],
+        'multipart'   => [],
+        'query'       => [],
+    ];
 
     /**
      * Response instance
      *
      * The response object will be set once the request has been made.
      *
-     * @var ResponseInterface
+     * @var ?ResponseInterface
      */
     protected $response;
 
@@ -74,17 +88,37 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
     protected $forceHttpMethod = false;
 
     /**
-     * {@inheritdoc}
+     * Request / response history for the Guzzle Client
+     *
+     * @var array{request: RequestInterface, response: ResponseInterface}[]
      */
-    public function setClient(ClientInterface $client) {
+    protected $clientHistory = [];
+
+    /**
+     * Error message used when a required response instance if missing
+     *
+     * @var string
+     */
+    protected $missingResponseError = 'The request has not been made yet, so no response object exists.';
+
+    /**
+     * Set the client instance
+     *
+     * @return self
+     */
+    public function setClient(ClientInterface $client, string $baseUri) {
         $this->client = $client;
-        $this->request = new Request('GET', $client->getConfig('base_uri'));
+        $this->baseUri = $baseUri;
+
+        $this->request = new Request('GET', $this->baseUri);
 
         return $this;
     }
 
     /**
-     * {@inheritdoc}
+     * Set the array contains comparator instance
+     *
+     * @return self
      */
     public function setArrayContainsComparator(ArrayContainsComparator $comparator) {
         $this->arrayContainsComparator = $comparator;
@@ -108,9 +142,15 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
             throw new InvalidArgumentException(sprintf('File does not exist: "%s"', $path));
         }
 
+        $contents = fopen($path, 'r');
+
+        if (false === $contents) {
+            throw new RuntimeException(sprintf('Unable to open file: %s', $path));
+        }
+
         return $this->addMultipartPart([
-            'name' => $partName,
-            'contents' => fopen($path, 'r'),
+            'name'     => $partName,
+            'contents' => $contents,
             'filename' => basename($path),
         ]);
     }
@@ -118,14 +158,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
     /**
      * Add an element to the multipart array
      *
-     * @param array $part The part to add
+     * @param array{name: string, contents: resource|string, filename?: string} $part The part to add
      * @return self
      */
     private function addMultipartPart($part) {
-        if (!isset($this->requestOptions['multipart'])) {
-            $this->requestOptions['multipart'] = [];
-        }
-
         $this->requestOptions['multipart'][] = $part;
 
         return $this;
@@ -140,10 +176,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @Given the following multipart form parameters are set:
      */
     public function setRequestMultipartFormParams(TableNode $table) {
-        foreach ($table as $row) {
+        foreach ($this->getTableNodeHash($table) as $name => $value) {
             $this->addMultipartPart([
-                'name' => $row['name'],
-                'contents' => $row['value'],
+                'name'     => $name,
+                'contents' => $value,
             ]);
         }
 
@@ -161,6 +197,67 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      */
     public function setBasicAuth($username, $password) {
         $this->requestOptions['auth'] = [$username, $password];
+
+        return $this;
+    }
+
+    /**
+     * Send OAuth request using password grant and set Authorization header.
+     *
+     * @param string $path         The path to get the token from
+     * @param string $username     The username to authenticate with
+     * @param string $password     The password to authenticate with
+     * @param string $scope        The scope to authenticate in
+     * @param string $clientId     The client_id to send
+     * @param string $clientSecret Optional client_secret to send
+     * @return self
+     *
+     * @Given I get an OAuth token using password grant from :path with :username and :password in scope :scope using client ID :clientId
+     * @Given I get an OAuth token using password grant from :path with :username and :password in scope :scope using client ID :clientId and client secret :clientSecret
+     */
+    public function oauthWithPasswordGrantInScope($path, $username, $password, $scope, $clientId, $clientSecret = null) {
+        $this->requestOptions['form_params'] = array_filter([
+            'grant_type'    => 'password',
+            'username'      => $username,
+            'password'      => $password,
+            'scope'         => $scope,
+            'client_id'     => $clientId,
+            'client_secret' => $clientSecret,
+        ]);
+
+        $this->addRequestHeader('Content-Type', 'application/x-www-form-urlencoded');
+        $this->setRequestPath($path);
+        $this->setRequestMethod('POST');
+        $this->sendRequest();
+
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
+
+        $this->requestOptions['form_params'] = [];
+
+        if (200 !== $this->response->getStatusCode()) {
+            throw new RuntimeException(sprintf(
+                'Expected request for access token to pass, got status code %d with the following response: %s',
+                $this->response->getStatusCode(),
+                (string) $this->response->getBody()
+            ));
+        }
+
+        /** @var stdClass */
+        $body = $this->getResponseBody();
+
+        if (empty($body->access_token)) {
+            throw new RuntimeException(sprintf(
+                'Missing access_token from response body: %s',
+                json_encode($body))
+            );
+        }
+
+        $this->addRequestHeader(
+            'Authorization',
+            sprintf('Bearer %s', (string) $body->access_token)
+        );
 
         return $this;
     }
@@ -208,20 +305,18 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @Given the following form parameters are set:
      */
     public function setRequestFormParams(TableNode $table) {
-        if (!isset($this->requestOptions['form_params'])) {
-            $this->requestOptions['form_params'] = [];
-        }
+        /** @var array<string, array{name: string, value: string}> */
+        $rows = $table->getColumnsHash();
 
-        foreach ($table as $row) {
-            $name = $row['name'];
+        foreach ($rows as $row) {
+            $name  = $row['name'];
             $value = $row['value'];
 
             if (isset($this->requestOptions['form_params'][$name]) && !is_array($this->requestOptions['form_params'][$name])) {
-                $this->requestOptions['form_params'][$name] = [$this->requestOptions['form_params'][$name]];
-            }
-
-            if (isset($this->requestOptions['form_params'][$name])) {
-                $this->requestOptions['form_params'][$name][] = $value;
+                $this->requestOptions['form_params'][$name] = [
+                    $this->requestOptions['form_params'][$name],
+                    $value
+                ];
             } else {
                 $this->requestOptions['form_params'][$name] = $value;
             }
@@ -247,7 +342,11 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
             );
         }
 
-        $this->request = $this->request->withBody(Psr7\stream_for($string));
+        if ($string instanceof PyStringNode) {
+            $string = (string) $string;
+        }
+
+        $this->request = $this->request->withBody(Utils::streamFor($string));
 
         return $this;
     }
@@ -261,7 +360,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * mime type of the file.
      *
      * @param string $path Path to a file
-     * @throws InvalidArgumentException
+     * @throws InvalidArgumentException|RuntimeException
      * @return self
      *
      * @Given the request body contains :path
@@ -275,10 +374,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
             throw new InvalidArgumentException(sprintf('File is not readable: "%s"', $path));
         }
 
+        /** @var resource */
+        $fp = fopen($path, 'r');
+
         // Set the Content-Type request header and the request body
         return $this
-            ->setRequestHeader('Content-Type', mime_content_type($path))
-            ->setRequestBody(fopen($path, 'r'));
+            ->setRequestHeader('Content-Type', (string) mime_content_type($path))
+            ->setRequestBody($fp);
     }
 
     /**
@@ -308,6 +410,43 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
     }
 
     /**
+     * Add a query parameter to the upcoming request
+     *
+     * @param string $name The name of the parameter
+     * @param string|TableNode $value The value to add
+     * @return self
+     *
+     * @Given the query parameter :name is :value
+     * @Given the query parameter :name is:
+     */
+    public function setQueryStringParameter($name, $value) {
+        if ($value instanceof TableNode) {
+            /** @var string[] */
+            $value = array_column($value->getHash(), 'value');
+        }
+
+        $this->requestOptions['query'][$name] = $value;
+
+        return $this;
+    }
+
+    /**
+     * Set multiple query parameters for the upcoming request
+     *
+     * @param TableNode $params The values to set
+     * @return self
+     *
+     * @Given the following query parameters are set:
+     */
+    public function setQueryStringParameters(TableNode $params) {
+        foreach ($this->getTableNodeHash($params) as $name => $value) {
+            $this->requestOptions['query'][$name] = $value;
+        }
+
+        return $this;
+    }
+
+    /**
      * Request a path
      *
      * @param string $path The path to request
@@ -332,47 +471,53 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
     /**
      * Assert the HTTP response code
      *
-     * @param int $code The HTTP response code
+     * @param int|string $code The HTTP response code
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response code is :code
      */
-    public function assertResponseCodeIs($code) {
-        $this->requireResponse();
+    public function assertResponseCodeIs($code) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::same(
                 $actual = $this->response->getStatusCode(),
-                $expected = $this->validateResponseCode($code),
+                $expected = $this->validateResponseCode((int) $code),
                 sprintf('Expected response code %d, got %d.', $expected, $actual)
             );
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
      * Assert the HTTP response code is not a specific code
      *
-     * @param int $code The HTTP response code
+     * @param int|string $code The HTTP response code
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response code is not :code
      */
-    public function assertResponseCodeIsNot($code) {
-        $this->requireResponse();
+    public function assertResponseCodeIsNot($code) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::notSame(
                 $actual = $this->response->getStatusCode(),
-                $this->validateResponseCode($code),
+                $this->validateResponseCode((int) $code),
                 sprintf('Did not expect response code %d.', $actual)
             );
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -380,12 +525,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $phrase Expected HTTP response reason phrase
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response reason phrase is :phrase
      */
-    public function assertResponseReasonPhraseIs($phrase) {
-        $this->requireResponse();
+    public function assertResponseReasonPhraseIs($phrase) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::same($phrase, $actual = $this->response->getReasonPhrase(), sprintf(
@@ -396,6 +542,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -403,12 +551,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $phrase Reason phrase that the HTTP response should not equal
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response reason phrase is not :phrase
      */
-    public function assertResponseReasonPhraseIsNot($phrase) {
-        $this->requireResponse();
+    public function assertResponseReasonPhraseIsNot($phrase) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::notSame($phrase, $this->response->getReasonPhrase(), sprintf(
@@ -418,6 +567,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -425,12 +576,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $pattern Regular expression pattern
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response reason phrase matches :expression
      */
-    public function assertResponseReasonPhraseMatches($pattern) {
-        $this->requireResponse();
+    public function assertResponseReasonPhraseMatches($pattern) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::regex(
@@ -444,6 +596,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -451,12 +605,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $line Expected HTTP response status line
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response status line is :line
      */
-    public function assertResponseStatusLineIs($line) {
-        $this->requireResponse();
+    public function assertResponseStatusLineIs($line) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             $actualStatusLine = sprintf(
@@ -473,6 +628,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -480,12 +637,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $line Value that the HTTP response status line must not equal
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response status line is not :line
      */
-    public function assertResponseStatusLineIsNot($line) {
-        $this->requireResponse();
+    public function assertResponseStatusLineIsNot($line) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             $actualStatusLine = sprintf(
@@ -501,6 +659,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -508,12 +668,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $pattern Regular expression pattern
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response status line matches :expression
      */
-    public function assertResponseStatusLineMatches($pattern) {
-        $this->requireResponse();
+    public function assertResponseStatusLineMatches($pattern) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             $actualStatusLine = sprintf(
@@ -533,6 +694,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -548,16 +711,19 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $group Name of the group that the response code should be in
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response is :group
      */
-    public function assertResponseIs($group) {
-        $this->requireResponse();
+    public function assertResponseIs($group) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
+
         $range = $this->getResponseCodeGroupRange($group);
+        $code  = $this->response->getStatusCode();
 
         try {
-            Assertion::range($code = $this->response->getStatusCode(), $range['min'], $range['max']);
+            Assertion::range($code, $range['min'], $range['max']);
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException(sprintf(
                 'Expected response group "%s", got "%s" (response code: %d).',
@@ -566,6 +732,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
                 $code
             ));
         }
+
+        return true;
     }
 
     /**
@@ -581,16 +749,19 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $group Name of the group that the response code is not in
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response is not :group
      */
-    public function assertResponseIsNot($group) {
+    public function assertResponseIsNot($group) : bool {
         try {
             $this->assertResponseIs($group);
         } catch (AssertionFailedException $e) {
             // As expected, return
-            return;
+            return true;
+        }
+
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
         }
 
         throw new AssertionFailedException(sprintf(
@@ -605,12 +776,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $header Then name of the header
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the :header response header exists
      */
-    public function assertResponseHeaderExists($header) {
-        $this->requireResponse();
+    public function assertResponseHeaderExists($header) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::true(
@@ -620,6 +792,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -627,12 +801,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $header Then name of the header
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the :header response header does not exist
      */
-    public function assertResponseHeaderDoesNotExist($header) {
-        $this->requireResponse();
+    public function assertResponseHeaderDoesNotExist($header) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::false(
@@ -642,6 +817,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -650,12 +827,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $header The name of the header
      * @param string $value The value to compare with
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the :header response header is :value
      */
-    public function assertResponseHeaderIs($header, $value) {
-        $this->requireResponse();
+    public function assertResponseHeaderIs($header, $value) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::same(
@@ -671,6 +849,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -679,12 +859,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $header The name of the header
      * @param string $value The value to compare with
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the :header response header is not :value
      */
-    public function assertResponseHeaderIsNot($header, $value) {
-        $this->requireResponse();
+    public function assertResponseHeaderIsNot($header, $value) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::notSame(
@@ -699,6 +880,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -707,12 +890,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $header The name of the header
      * @param string $pattern The regular expression pattern
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the :header response header matches :pattern
      */
-    public function assertResponseHeaderMatches($header, $pattern) {
-        $this->requireResponse();
+    public function assertResponseHeaderMatches($header, $pattern) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
 
         try {
             Assertion::regex(
@@ -728,22 +912,23 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
      * Assert that the response body contains an empty JSON object
      *
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is an empty JSON object
      */
-    public function assertResponseBodyIsAnEmptyJsonObject() {
+    public function assertResponseBodyIsAnEmptyJsonObject() : bool {
         $this->requireResponse();
         $body = $this->getResponseBody();
 
         try {
-            Assertion::isInstanceOf($body, 'stdClass', 'Expected response body to be a JSON object.');
+            Assertion::isInstanceOf($body, stdClass::class, 'Expected response body to be a JSON object.');
             Assertion::same('{}', $encoded = json_encode($body, JSON_PRETTY_PRINT), sprintf(
                 'Expected response body to be an empty JSON object, got "%s".',
                 $encoded
@@ -751,17 +936,18 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
      * Assert that the response body contains an empty JSON array
      *
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is an empty JSON array
      */
-    public function assertResponseBodyIsAnEmptyJsonArray() {
+    public function assertResponseBodyIsAnEmptyJsonArray() : bool {
         $this->requireResponse();
 
         try {
@@ -773,18 +959,19 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
      * Assert that the response body contains an array with a specific length
      *
-     * @param int $length The length of the array
+     * @param int|string $length The length of the array
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is a JSON array of length :length
      */
-    public function assertResponseBodyJsonArrayLength($length) {
+    public function assertResponseBodyJsonArrayLength($length) : bool {
         $this->requireResponse();
         $length = (int) $length;
 
@@ -803,18 +990,19 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
      * Assert that the response body contains an array with a length of at least a given length
      *
-     * @param int $length The length to use in the assertion
+     * @param int|string $length The length to use in the assertion
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is a JSON array with a length of at least :length
      */
-    public function assertResponseBodyJsonArrayMinLength($length) {
+    public function assertResponseBodyJsonArrayMinLength($length) : bool {
         $this->requireResponse();
 
         $length = (int) $length;
@@ -827,7 +1015,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
                 sprintf(
                     'Expected response body to be a JSON array with at least %d entr%s, got %d: "%s".',
                     $length,
-                    (int) $length === 1 ? 'y' : 'ies',
+                    $length === 1 ? 'y' : 'ies',
                     $bodyLength,
                     json_encode($body, JSON_PRETTY_PRINT)
                 )
@@ -835,18 +1023,19 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
      * Assert that the response body contains an array with a length of at most a given length
      *
-     * @param int $length The length to use in the assertion
+     * @param int|string $length The length to use in the assertion
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is a JSON array with a length of at most :length
      */
-    public function assertResponseBodyJsonArrayMaxLength($length) {
+    public function assertResponseBodyJsonArrayMaxLength($length) : bool {
         $this->requireResponse();
 
         $length = (int) $length;
@@ -859,7 +1048,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
                 sprintf(
                     'Expected response body to be a JSON array with at most %d entr%s, got %d: "%s".',
                     $length,
-                    (int) $length === 1 ? 'y' : 'ies',
+                    $length === 1 ? 'y' : 'ies',
                     $bodyLength,
                     json_encode($body, JSON_PRETTY_PRINT)
                 )
@@ -867,6 +1056,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
 
@@ -875,12 +1066,14 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param PyStringNode $content The content to match the response body against
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is:
      */
-    public function assertResponseBodyIs(PyStringNode $content) {
-        $this->requireResponse();
+    public function assertResponseBodyIs(PyStringNode $content) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
+
         $content = (string) $content;
 
         try {
@@ -892,6 +1085,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -899,12 +1094,14 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param PyStringNode $content The content that the response body should not match
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body is not:
      */
-    public function assertResponseBodyIsNot(PyStringNode $content) {
-        $this->requireResponse();
+    public function assertResponseBodyIsNot(PyStringNode $content) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
+
         $content = (string) $content;
 
         try {
@@ -915,6 +1112,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -922,12 +1121,14 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param PyStringNode $pattern The regular expression pattern to use for the match
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body matches:
      */
-    public function assertResponseBodyMatches(PyStringNode $pattern) {
-        $this->requireResponse();
+    public function assertResponseBodyMatches(PyStringNode $pattern) : bool {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
+
         $pattern = (string) $pattern;
 
         try {
@@ -939,6 +1140,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         } catch (AssertionFailure $e) {
             throw new AssertionFailedException($e->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -946,18 +1149,18 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param PyStringNode $contains
      * @throws AssertionFailedException
-     * @return void
      *
      * @Then the response body contains JSON:
      */
-    public function assertResponseBodyContainsJson(PyStringNode $contains) {
+    public function assertResponseBodyContainsJson(PyStringNode $contains) : bool {
         $this->requireResponse();
 
         // Decode the parameter to the step as an array and make sure it's valid JSON
         $contains = $this->jsonDecode((string) $contains);
 
         // Get the decoded response body and make sure it's decoded to an array
-        $body = json_decode(json_encode($this->getResponseBody()), true);
+        /** @var array<array-key, mixed> */
+        $body = json_decode((string) json_encode($this->getResponseBody()), true);
 
         try {
             // Compare the arrays, on error this will throw an exception
@@ -967,6 +1170,8 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
                 'Comparator did not return in a correct manner. Marking assertion as failed.'
             );
         }
+
+        return true;
     }
 
     /**
@@ -1004,14 +1209,13 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
                 }
             }
 
-            // Remove form_params from the options, otherwise Guzzle will throw an exception
-            unset($this->requestOptions['form_params']);
+            $this->requestOptions['form_params'] = [];
         }
 
         try {
             $this->response = $this->client->send(
                 $this->request,
-                $this->requestOptions
+                array_filter($this->requestOptions)
             );
         } catch (RequestException $e) {
             $this->response = $e->getResponse();
@@ -1029,7 +1233,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @throws RuntimeException
      */
-    protected function requireResponse() {
+    protected function requireResponse() : void {
         if (!$this->response) {
             throw new RuntimeException('The request has not been made yet, so no response object exists.');
         }
@@ -1040,10 +1244,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      *
      * @param string $group The name of the group
      * @throws InvalidArgumentException
-     * @return array An array with two keys, min and max, which represents the min and max values
-     *               for $group
+     * @return array{min: int, max: int} An array with two keys, min and max, which represents the
+     *                                   min and max values for $group
      */
-    protected function getResponseCodeGroupRange($group) {
+    protected function getResponseCodeGroupRange(string $group) : array {
         switch ($group) {
             case 'informational':
                 $min = 100;
@@ -1081,9 +1285,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param int $code The respose code
      * @return string
      */
-    protected function getResponseGroup($code) {
-        $code = (int) $code;
-
+    protected function getResponseGroup(int $code) : string {
         if ($code >= 500) {
             return 'server error';
         } else if ($code >= 400) {
@@ -1104,9 +1306,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @throws InvalidArgumentException
      * @return int
      */
-    protected function validateResponseCode($code) {
-        $code = (int) $code;
-
+    protected function validateResponseCode(int $code) : int {
         try {
             Assertion::range($code, 100, 599, sprintf('Response code must be between 100 and 599, got %d.', $code));
         } catch (AssertionFailure $e) {
@@ -1122,9 +1322,9 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $path The path to request
      * @return self
      */
-    protected function setRequestPath($path) {
-        // Resolve the path with the base_uri set in the client
-        $uri = Psr7\Uri::resolve($this->client->getConfig('base_uri'), Psr7\uri_for($path));
+    protected function setRequestPath(string $path) {
+        $base = Utils::uriFor($this->baseUri);
+        $uri = UriResolver::resolve($base, Utils::uriFor($path));
         $this->request = $this->request->withUri($uri);
 
         return $this;
@@ -1134,12 +1334,12 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Update the HTTP method of the request
      *
      * @param string $method The HTTP method
-     * @param boolean $force Force the HTTP method. If set to false the method set CAN be
+     * @param bool $force Force the HTTP method. If set to false the method set CAN be
      *                       overridden (this occurs for instance when adding form parameters to the
      *                       request, and not specifying HTTP POST for the request)
      * @return self
      */
-    protected function setRequestMethod($method, $force = true) {
+    protected function setRequestMethod(string $method, bool $force = true) {
         $this->request = $this->request->withMethod($method);
         $this->forceHttpMethod = $force;
 
@@ -1150,9 +1350,14 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * Get the JSON-encoded array or stdClass from the response body
      *
      * @throws InvalidArgumentException
-     * @return array|stdClass
+     * @return array<mixed>|stdClass
      */
     protected function getResponseBody() {
+        if (!$this->response) {
+            throw new RuntimeException($this->missingResponseError);
+        }
+
+        /** @var mixed */
         $body = json_decode((string) $this->response->getBody());
 
         if (json_last_error() !== JSON_ERROR_NONE) {
@@ -1161,6 +1366,7 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
             throw new InvalidArgumentException('The response body does not contain a valid JSON array / object.');
         }
 
+        /** @var array<mixed>|stdClass */
         return $body;
     }
 
@@ -1170,8 +1376,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @throws InvalidArgumentException
      * @return array
      */
-    protected function getResponseBodyArray() {
-        if (!is_array($body = $this->getResponseBody())) {
+    protected function getResponseBodyArray() : array {
+        $body = $this->getResponseBody();
+
+        if (!is_array($body)) {
             throw new InvalidArgumentException('The response body does not contain a valid JSON array.');
         }
 
@@ -1184,9 +1392,10 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
      * @param string $value The value to decode
      * @param string $errorMessage Optional error message
      * @throws InvalidArgumentException
-     * @return array
+     * @return array<mixed>
      */
-    protected function jsonDecode($value, $errorMessage = null) {
+    protected function jsonDecode(string $value, string $errorMessage = null) : array {
+        /** @var array<mixed> */
         $decoded = json_decode($value, true);
 
         if (json_last_error() !== JSON_ERROR_NONE) {
@@ -1196,5 +1405,18 @@ class ApiContext implements ApiClientAwareContext, ArrayContainsComparatorAwareC
         }
 
         return $decoded;
+    }
+
+    /**
+     * Get an associative array from the TableNode
+     *
+     * This method will effectively remove duplicates from TableNode
+     *
+     * @param TableNode $table
+     * @return array<string, string>
+     */
+    protected function getTableNodeHash(TableNode $table) : array {
+        /** @var array<string, string> */
+        return array_slice($table->getRowsHash(), 1);
     }
 }
