@@ -1,31 +1,24 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
+use InvalidArgumentException;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\ArrayMaxLength
- * @testdox Array max length matcher
  */
-class ArrayMaxLengthTest extends PHPUnit_Framework_TestCase {
-    /**
-     * @var ArrayMaxLength
-     */
+class ArrayMaxLengthTest extends TestCase {
+    /** @var ArrayMaxLength */
     private $matcher;
 
-    /**
-     * Set up matcher instance
-     */
-    public function setup() {
+    public function setup() : void {
         $this->matcher = new ArrayMaxLength();
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array{list: int[], length: int}[]
      */
-    public function getArraysAndLengths() {
+    public function getArraysAndLengths() : array {
         return [
             [
                 'list' => [],
@@ -43,11 +36,9 @@ class ArrayMaxLengthTest extends PHPUnit_Framework_TestCase {
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array{value: int|string|array<string, string>, message: string}[]
      */
-    public function getInvalidValues() {
+    public function getInvalidValues() : array {
         return [
             [
                 'value' => 123,
@@ -65,11 +56,9 @@ class ArrayMaxLengthTest extends PHPUnit_Framework_TestCase {
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array{array: int[], maxLength: int, message: string}[]
      */
-    public function getValuesThatFail() {
+    public function getValuesThatFail() : array {
         return [
             [
                 'array' => [1, 2],
@@ -87,27 +76,23 @@ class ArrayMaxLengthTest extends PHPUnit_Framework_TestCase {
     /**
      * @dataProvider getArraysAndLengths
      * @covers ::__invoke
-     *
-     * @param array $array
-     * @param int $length
+     * @param int[] $array
      */
-    public function testCanMatchMaxLengthOfArrays(array $array, $length) {
+    public function testCanMatchMaxLengthOfArrays(array $array, int $length) : void {
         $matcher = $this->matcher;
-        $this->assertNull(
+        $this->assertTrue(
             $matcher($array, $length),
-            'Matcher is supposed to return null.'
+            'Matcher is supposed to return true.'
         );
     }
 
     /**
      * @dataProvider getInvalidValues
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     *
-     * @param mixed $value
-     * @param string $message
+     * @param int|string|array<string, string> $value
      */
-    public function testThrowsExceptionWhenMatchingAgainstAnythingOtherThanAnArray($value, $message) {
+    public function testThrowsExceptionWhenMatchingAgainstAnythingOtherThanAnArray($value, string $message) : void {
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage($message);
         $matcher = $this->matcher;
         $matcher($value, 123);
@@ -116,13 +101,10 @@ class ArrayMaxLengthTest extends PHPUnit_Framework_TestCase {
     /**
      * @dataProvider getValuesThatFail
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     *
-     * @param array $array
-     * @param int $maxLength
-     * @param string $message
+     * @param int[] $array
      */
-    public function testThrowsExceptionWhenLengthIsTooShort(array $array, $maxLength, $message) {
+    public function testThrowsExceptionWhenLengthIsTooShort(array $array, int $maxLength, string $message) : void {
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage($message);
         $matcher = $this->matcher;
         $matcher($array, $maxLength);
