@@ -1,31 +1,24 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
-use PHPUnit_Framework_TestCase;
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\GreaterThan
- * @testdox Numeric greater than matcher
  */
-class GreaterThanTest extends PHPUnit_Framework_TestCase {
-    /**
-     * @var GreaterThan
-     */
+class GreaterThanTest extends TestCase {
+    /** @var GreaterThan */
     private $matcher;
 
-    /**
-     * Set up matcher instance
-     */
-    public function setup() {
+    public function setUp() : void {
         $this->matcher = new GreaterThan();
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array{number: int|float|string, min: int|float|string}[]
      */
-    public function getValuesForMatching() {
+    public function getValuesForMatching() : array {
         return [
             'integer' => [
                 'number' => 2,
@@ -43,11 +36,9 @@ class GreaterThanTest extends PHPUnit_Framework_TestCase {
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array{number: int|string|float, min: int|string|float, errorMessage: string}[]
      */
-    public function getFailingValues() {
+    public function getFailingValues() : array {
         return [
             [
                 'number' => 123,
@@ -75,50 +66,47 @@ class GreaterThanTest extends PHPUnit_Framework_TestCase {
     /**
      * @dataProvider getValuesForMatching
      * @covers ::__invoke
-     *
-     * @param numeric $number
-     * @param numeric $min
+     * @param int|float|string $number
+     * @param int|float|string $min
      */
-    public function testCanCompareValuesOfType($number, $min) {
+    public function testCanCompareValuesOfType($number, $min) : void {
         $matcher = $this->matcher;
-        $this->assertNull(
+        $this->assertTrue(
             $matcher($number, $min),
-            'Matcher is supposed to return null.'
+            'Matcher is supposed to return true.'
         );
     }
 
     /**
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage "foo" is not numeric.
      */
-    public function testThrowsExceptionIfNumberIsNotNumeric() {
+    public function testThrowsExceptionIfNumberIsNotNumeric() : void {
         $matcher = $this->matcher;
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('"foo" is not numeric.');
         $matcher('foo', 123);
     }
 
     /**
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage "foo" is not numeric.
      */
-    public function testThrowsExceptionIfMinimumNumberIsNotNumeric() {
+    public function testThrowsExceptionIfMinimumNumberIsNotNumeric() : void {
         $matcher = $this->matcher;
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('"foo" is not numeric.');
         $matcher(123, 'foo');
     }
 
     /**
      * @dataProvider getFailingValues
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     *
-     * @param numeric $number
-     * @param numeric $min
-     * @param string $errorMessage
+     * @param int|string|float $number
+     * @param int|string|float $min
      */
-    public function testThrowsExceptionWhenComparisonFails($number, $min, $errorMessage) {
-        $this->expectExceptionMessage($errorMessage);
+    public function testThrowsExceptionWhenComparisonFails($number, $min, string $errorMessage) : void {
         $matcher = $this->matcher;
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage($errorMessage);
         $matcher($number, $min);
     }
 }
