@@ -1,31 +1,24 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
+use InvalidArgumentException;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\VariableType
- * @testdox Variable type matcher
  */
-class VariableTypeTest extends PHPUnit_Framework_TestCase {
-    /**
-     * @var VariableType
-     */
+class VariableTypeTest extends TestCase {
+    /** @var VariableType */
     private $matcher;
 
-    /**
-     * Set up matcher instance
-     */
-    public function setup() {
+    public function setUp() : void {
         $this->matcher = new VariableType();
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array<string, array{value: mixed, type: string}>
      */
-    public function getValuesAndTypes() {
+    public function getValuesAndTypes() : array {
         return [
             'int' => [
                 'value' => 1,
@@ -95,15 +88,49 @@ class VariableTypeTest extends PHPUnit_Framework_TestCase {
                 'value' => ['foo' => 'bar'],
                 'type' => 'object',
             ],
+            'boolean (any)' => [
+                'value' => true,
+                'type' => 'any',
+            ],
+            'integer (any)' => [
+                'value' => 123,
+                'type' => 'any',
+            ],
+            'double (any)' => [
+                'value' => 1.1,
+                'type' => 'any',
+            ],
+            'string (any)' => [
+                'value' => 'some string',
+                'type' => 'any',
+            ],
+            'array (any)' => [
+                'value' => [1, 2, 3],
+                'type' => 'any',
+            ],
+            'object (any)' => [
+                'value' => ['foo' => 'bar'],
+                'type' => 'any',
+            ],
+            'int (multiple)' => [
+                'value' => 1,
+                'type' => 'string|array|integer',
+            ],
+            'integer (multiple)' => [
+                'value' => 1,
+                'type' => 'int|bool|double',
+            ],
+            'string (multiple)' => [
+                'value' => 'some string',
+                'type' => 'integer | bool | array | string', // spaces are intentional
+            ],
         ];
     }
 
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array{value: mixed, type: string, message: string}[]
      */
-    public function getInvalidMatches() {
+    public function getInvalidMatches() : array {
         return [
             [
                 'value' => 123,
@@ -131,41 +158,36 @@ class VariableTypeTest extends PHPUnit_Framework_TestCase {
     /**
      * @dataProvider getValuesAndTypes
      * @covers ::__invoke
-     * @covers ::normalizeType
-     *
+     * @covers ::normalizeTypes
      * @param mixed $value
-     * @param string $type
      */
-    public function testCanMatchValuesOfType($value, $type) {
+    public function testCanMatchValuesOfType($value, string $type) : void {
         $matcher = $this->matcher;
-        $this->assertNull(
+        $this->assertTrue(
             $matcher($value, $type),
-            'Matcher is supposed to return null.'
+            'Matcher is supposed to return true.'
         );
     }
 
     /**
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage Unsupported variable type: "resource".
      */
-    public function testThrowsExceptionWhenGivenInvalidType() {
+    public function testThrowsExceptionWhenGivenInvalidType() : void {
         $matcher = $this->matcher;
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Unsupported variable type: "resource".');
         $matcher('foo', 'resource');
     }
 
     /**
      * @dataProvider getInvalidMatches
      * @covers ::__invoke
-     * @expectedException InvalidArgumentException
-     *
      * @param mixed $value
-     * @param string $type
-     * @param string $message
      */
-    public function testThrowsExceptionWhenTypeOfValueDoesNotMatchExpectedType($value, $type, $message) {
-        $this->expectExceptionMessage($message);
+    public function testThrowsExceptionWhenTypeOfValueDoesNotMatchExpectedType($value, string $type, string $message) : void {
         $matcher = $this->matcher;
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage($message);
         $matcher($value, $type);
     }
 }
