@@ -1,37 +1,31 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\Context\Initializer;
 
 use Imbo\BehatApiExtension\Context\ApiClientAwareContext;
 use GuzzleHttp\Client;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
+use RuntimeException;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\Context\Initializer\ApiClientAwareInitializer
- * @testdox Initializer for API Client aware contexts
  */
-class ApiClientAwareInitializerTest extends PHPUnit_Framework_TestCase {
+class ApiClientAwareInitializerTest extends TestCase {
     /**
      * @covers ::initializeContext
-     * @covers ::validateConnection
-     * @expectedException RuntimeException
-     * @expectedExceptionMessage Can't connect to base_uri: "http://localhost:123".
-     */
-    public function testThrowsExceptionWhenBaseUriIsNotConnectable() {
-        $initializer = new ApiClientAwareInitializer(['base_uri' => 'http://localhost:123']);
-        $initializer->initializeContext($this->createMock(ApiClientAwareContext::class));
-    }
-
-    /**
-     * @covers ::initializeContext
-     * @covers ::validateConnection
      * @covers ::__construct
      */
-    public function testInjectsClientWhenInitializingContext() {
+    public function testInjectsClientWhenInitializingContext() : void {
         // Set up a socket for the test case, try all ports between 8000 and 8079. If no ports are
         // available the test case will be marked as skipped. This is to get past the base URI
         // validation
         set_error_handler(function() { return true; });
         $sock = socket_create(AF_INET, SOCK_STREAM, SOL_TCP);
+
+        if (false === $sock) {
+            $this->fail('Unable to create socket');
+        }
+
+        $result = false;
 
         for ($port = 8000; $port < 8079; $port++) {
             if ($result = socket_bind($sock, 'localhost', $port)) {
@@ -47,7 +41,9 @@ class ApiClientAwareInitializerTest extends PHPUnit_Framework_TestCase {
         }
 
         // Listen for connections
-        socket_listen($sock);
+        if (!socket_listen($sock)) {
+            $this->markTestSkipped('Unable to listen for a connection, skipping test for now.');
+        }
 
         $context = $this->createMock(ApiClientAwareContext::class);
         $context
