@@ -1,19 +1,17 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\Exception;
 
-use PHPUnit_Framework_TestCase;
+use Imbo\BehatApiExtension\Exception\ArrayContainsComparatorException;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @coversDefaultClass Imbo\BehatApiExtension\Exception\ArrayContainsComparatorException
- * @testdox Array contains comparator exception
  */
-class ArrayContainsComparatorExceptionTest extends PHPUnit_Framework_TestCase {
+class ArrayContainsComparatorExceptionTest extends TestCase {
     /**
-     * Data provider
-     *
-     * @return array[]
+     * @return array<array-key, array{message: string, needle: array<string, string>, haystack: array<string, string>, formattedMessage: string}>
      */
-    public function getExceptionData() {
+    public function getExceptionData() : array {
         return [
             'with no needle / haystack' => [
                 'message' => $someMessage = 'some message',
@@ -35,8 +33,8 @@ MESSAGE
             ],
             'with needle and haystack' => [
                 'message' => $someMessage = 'some message',
-                'needle' => $needle = ['needle' => 'value'],
-                'haystack' => $haystack = ['haystack' => 'value'],
+                'needle' => ['needle' => 'value'],
+                'haystack' => ['haystack' => 'value'],
                 'formattedMessage' => <<<MESSAGE
 {$someMessage}
 
@@ -60,15 +58,14 @@ MESSAGE
 
     /**
      * @dataProvider getExceptionData
-     * @expectedException Imbo\BehatApiExtension\Exception\ArrayContainsComparatorException
      * @covers ::__construct
-     *
      * @param string $message
-     * @param array $needle
-     * @param array $haystack
+     * @param array<string, string> $needle
+     * @param array<string, string> $haystack
      * @param string $formattedMessage
      */
-    public function testCanProperlyFormatErrorMessages($message, array $needle, array $haystack, $formattedMessage) {
+    public function testCanProperlyFormatErrorMessages(string $message, array $needle, array $haystack, string $formattedMessage) : void {
+        $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage($formattedMessage);
         throw new ArrayContainsComparatorException($message, 0, null, $needle, $haystack);
     }
