@@ -53,9 +53,9 @@ copyright = u'2016-' + str(date.today().year) + ', Christer Edvartsen'
 # built documents.
 #
 # The short X.Y version.
-version = '2.2'
+version = '3.0'
 # The full version, including alpha/beta/rc tags.
-release = '2.2.1'
+release = '3.0.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
