@@ -1,5 +1,5 @@
 <?php
-use Behat\Behat\Context\SnippetAcceptingContext;
+use Behat\Behat\Context\Context;
 use Behat\Gherkin\Node\PyStringNode;
 use Behat\Testwork\Hook\Scope\SuiteScope;
 use Behat\Behat\Hook\Scope\BeforeScenarioScope;
@@ -15,25 +15,25 @@ use Assert\Assertion;
  *
  * @author Christer Edvartsen <cogo@starzinger.net>
  */
-class FeatureContext implements SnippetAcceptingContext {
+class FeatureContext implements Context {
     /**
      * PHP binary used to trigger Behat from the scenarios
      *
-     * @var string
+     * @var ?string
      */
     private $phpBin;
 
     /**
      * Process instance for executing processes
      *
-     * @var Process
+     * @var ?Process
      */
     private $process;
 
     /**
      * The working directory where files can be created
      *
-     * @var string
+     * @var ?string
      */
     private $workingDir;
 
@@ -45,7 +45,7 @@ class FeatureContext implements SnippetAcceptingContext {
      * @BeforeSuite
      * @AfterSuite
      */
-    public static function emptyTestDir(SuiteScope $scope) {
+    public static function emptyTestDir(SuiteScope $scope) : void {
         $testDir = sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'behat-api-extension';
 
         if (is_dir($testDir)) {
@@ -61,7 +61,7 @@ class FeatureContext implements SnippetAcceptingContext {
      *
      * @BeforeScenario
      */
-    public function prepareScenario(BeforeScenarioScope $scope) {
+    public function prepareScenario(BeforeScenarioScope $scope) : void {
         $dir = sys_get_temp_dir() . DIRECTORY_SEPARATOR . 'behat-api-extension' . DIRECTORY_SEPARATOR . microtime(true);
         mkdir($dir . '/features/bootstrap', 0777, true);
 
@@ -72,7 +72,6 @@ class FeatureContext implements SnippetAcceptingContext {
 
         $this->workingDir = $dir;
         $this->phpBin = $bin;
-        $this->process = new Process(null);
     }
 
     /**
@@ -84,8 +83,8 @@ class FeatureContext implements SnippetAcceptingContext {
      *
      * @Given a file named :filename with:
      */
-    public function createFile($filename, PyStringNode $content, $readable = true) {
-        $filename = rtrim($this->workingDir, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR . ltrim($filename, DIRECTORY_SEPARATOR);
+    public function createFile($filename, PyStringNode $content, $readable = true) : void {
+        $filename = rtrim((string) $this->workingDir, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR . ltrim($filename, DIRECTORY_SEPARATOR);
         $path = dirname($filename);
         $content = str_replace("'''", '"""', (string) $content);
 
@@ -108,29 +107,35 @@ class FeatureContext implements SnippetAcceptingContext {
      *
      * @Given a non-readable file named :filename with:
      */
-    public function createNonReadableFile($filename, PyStringNode $content) {
+    public function createNonReadableFile($filename, PyStringNode $content) : void {
         $this->createFile($filename, $content, false);
     }
 
     /**
      * Runs Behat
      *
-     * @param string $argumentsString
+     * @param string $args
      * @When /^I run "behat(?: ((?:\"|[^"])*))?"$/
      */
-    public function runBehat($args = '') {
+    public function runBehat(string $args = '') : void {
+        if (!defined('BEHAT_BIN_PATH')) {
+            throw new RuntimeException('Missing BEHAT_BIN_PATH constant');
+        }
+
         $args = strtr($args, ['\'' => '"']);
 
-        $this->process->setWorkingDirectory($this->workingDir);
-        $this->process->setCommandLine(
+        $this->process = Process::fromShellCommandline(
             sprintf(
                 '%s %s %s %s',
-                $this->phpBin,
-                escapeshellarg(BEHAT_BIN_PATH),
+                (string) $this->phpBin,
+                escapeshellarg((string) BEHAT_BIN_PATH),
                 $args,
                 '--format-settings="{\"timer\": false}" --no-colors'
-            )
+            ),
+            $this->workingDir
         );
+
+
         $this->process->start();
         $this->process->wait();
     }
@@ -143,7 +148,7 @@ class FeatureContext implements SnippetAcceptingContext {
      *
      * @Then /^it should (fail|pass) with:$/
      */
-    public function assertCommandResultWithOutput($result, PyStringNode $output) {
+    public function assertCommandResultWithOutput($result, PyStringNode $output) : void {
         $this->assertCommandResult($result);
         $this->assertCommandOutputMatches($output);
     }
@@ -151,11 +156,11 @@ class FeatureContext implements SnippetAcceptingContext {
     /**
      * Assert command output contains a string
      *
-     * @param PyStringNode $output
+     * @param PyStringNode $content
      *
      * @Then the output should contain:
      */
-    public function assertCommandOutputMatches(PyStringNode $content) {
+    public function assertCommandOutputMatches(PyStringNode $content) : void {
         Assertion::contains(
             $this->getOutput(),
             str_replace("'''", '"""', (string) $content),
@@ -169,7 +174,7 @@ class FeatureContext implements SnippetAcceptingContext {
      * @param string $result
      * @Then /^it should (fail|pass)$/
      */
-    public function assertCommandResult($result) {
+    public function assertCommandResult(string $result) : void {
         $exitCode = $this->getExitCode();
 
         // Escape % as the callback will pass this value to sprintf() if the assertion fails, and
@@ -200,8 +205,18 @@ class FeatureContext implements SnippetAcceptingContext {
      *
      * @return int
      */
-    private function getExitCode() {
-        return $this->process->getExitCode();
+    private function getExitCode() : int {
+        if (null === $this->process) {
+            throw new RuntimeException('No process is running');
+        }
+
+        $code = $this->process->getExitCode();
+
+        if (null === $code) {
+            throw new RuntimeException('Process is not finished');
+        }
+
+        return $code;
     }
 
     /**
@@ -210,9 +225,13 @@ class FeatureContext implements SnippetAcceptingContext {
      * @return string
      */
     private function getOutput() {
+        if (null === $this->process) {
+            throw new RuntimeException('No process is running');
+        }
+
         $output = $this->process->getErrorOutput() . $this->process->getOutput();
 
-        return trim(preg_replace('/ +$/m', '', $output));
+        return trim((string) preg_replace('/ +$/m', '', $output));
     }
 
     /**
@@ -220,8 +239,11 @@ class FeatureContext implements SnippetAcceptingContext {
      *
      * @param string $path Path to a file or a directory
      */
-    private static function rmdir($path) {
-        foreach (glob($path . '/*') as $file) {
+    private static function rmdir($path) : void {
+        /** @var string[] */
+        $files = glob(sprintf('%s/*', $path));
+
+        foreach ($files as $file) {
             if (is_dir($file)) {
                 self::rmdir($file);
             } else {
