@@ -1,44 +1,28 @@
 <?php
-/**
- * Silex application used to test the Behat API extension
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
- */
+namespace Imbo\BehatApiExtension;
 
-use Silex\Application;
-use Silex\Provider\SecurityServiceProvider;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Request;
+use Slim\Factory\AppFactory;
+use Tuupola\Middleware\HttpBasicAuthentication;
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
+use stdClass;
 
 require_once __DIR__ . '/../../vendor/autoload.php';
 
-$app = new Application();
-
-$app->register(new SecurityServiceProvider(), [
-    'security.firewalls' => [
-        'basicAuth' => [
-            'pattern' => '^/basicAuth',
-            'http' => true,
-            'users' => [
-                'foo' => [
-                    'ROLE_ADMIN',
-                    '$2y$13$jUzpuB1A0C0A5utO0hW/1eKngeCIxR8LO/Ios5Tay9b8zRUcLtCMO', // bar
-                ],
-                'bar' => [
-                    'ROLE_ADMIN',
-                    '$2y$10$3i9/lVd8UOFIJ6PAMFt8gu3/r5g0qeCJvoSlLCsvMTythye19F77a', // foo
-                ],
-            ],
-        ],
-    ],
-]);
+$app = AppFactory::create();
+$app->add(new HttpBasicAuthentication([
+    'path' => '/basicAuth',
+    'realm' => 'Protected',
+    'users' => [
+        'foo' => 'bar',
+    ]
+]));
 
 /**
  * Front page
  */
-$app->match('/', function(Request $request) {
-    return new JsonResponse([
+$app->any('/', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([
         'null' => null,
         'string' => 'value',
         'integer' => 42,
@@ -63,85 +47,153 @@ $app->match('/', function(Request $request) {
             'null' => null,
             'scalar' => '123',
         ],
-    ], 200, [
-        'X-Foo' => 'foo',
-    ]);
+    ]));
+    return $response
+        ->withHeader('Content-Type', 'application/json')
+        ->withHeader('X-Foo', 'foo');
+});
+
+/**
+ * List with objects
+ */
+$app->any('/list', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([
+        [
+            'integer' => 123,
+            'string' => 'value',
+        ]
+    ]));
+
+    return $response
+        ->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Echo the request body
  */
-$app->match('/echo', function(Request $request) {
-    $headers = [];
-
+$app->any('/echo', function(Request $request, Response $response) : Response {
     // Set the same Content-Type header in the response as found in the request
-    if ($contentType = $request->headers->get('Content-Type')) {
-        $headers['Content-Type'] = $contentType;
+    if ($contentType = $request->getHeaderLine('Content-Type')) {
+        $response = $response->withHeader('Content-Type', $contentType);
     }
 
-    return $request->query->has('json') ?
-        new JsonResponse(json_decode($request->getContent(), true), 200, $headers) :
-        new Response($request->getContent(), 200, $headers);
+    $requestBody = (string) $request->getBody();
+
+    if (array_key_exists('json', $request->getQueryParams())) {
+        $response->getBody()->write((string) json_encode(json_decode($requestBody, true)));
+        $response = $response->withHeader('Content-Type', 'application/json');
+    } else {
+        $response->getBody()->write($requestBody);
+    }
+
+    return $response;
 });
 
 /**
  * Return information about uploaded files
  */
-$app->post('/files', function(Request $request) {
-    return new JsonResponse($_FILES);
+$app->post('/files', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode($_FILES));
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return information about the request
  */
-$app->match('/requestInfo', function(Request $request) {
-    return new JsonResponse([
+$app->any('/requestInfo', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([
         '_GET' => $_GET,
         '_POST' => $_POST,
         '_FILES' => $_FILES,
         '_SERVER' => $_SERVER,
-        'requestBody' => $request->getContent(),
-    ]);
+        'requestBody' => (string) $request->getBody(),
+    ]));
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return the HTTP method
  */
-$app->match('/echoHttpMethod', function(Request $request) {
-    return new JsonResponse([
+$app->any('/echoHttpMethod', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([
         'method' => $request->getMethod(),
-    ]);
+    ]));
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return the authenticated user name
  */
-$app->match('/basicAuth', function(Application $app) {
-    return new JsonResponse([
-        'user' => $app['security.token_storage']->getToken()->getUser()->getUsername(),
-    ]);
+$app->any('/basicAuth', function(Request $request, Response $response) {
+    $response->getBody()->write((string) json_encode([
+        'user' => explode(':', $request->getUri()->getUserInfo())[0],
+    ]));
+
+    return $response->withHeader('Content-Type', 'application/json');
+});
+
+/**
+ * Return access token given the correct credentials
+ */
+$app->any('/oauth/token', function(Request $request, Response $response) {
+    /** @var array{username: string, password: string} */
+    $body = $request->getParsedBody();
+
+    if ('foo' === $body['username'] && 'bar' === $body['password']) {
+        $responseBody = [
+            'access_token' => 'some_access_token',
+        ];
+    } else {
+        $response = $response->withStatus(401);
+        $responseBody = [
+            'error' => 'invalid_request',
+        ];
+    }
+
+    $response->getBody()->write((string) json_encode($responseBody));
+
+    return $response->withHeader('Content-Type', 'application/json');
+});
+
+/**
+ * Return secured resource if Authorization header is valid.
+ */
+$app->any('/securedWithOAuth', function(Request $request, Response $response) {
+    if ('Bearer some_access_token' === $request->getHeaderLine('Authorization')) {
+        $responseBody = [
+            'users' => [
+                'foo' => 'bar',
+            ]
+        ];
+    } else {
+        $response = $response->withStatus(401);
+        $responseBody = [
+            'error' => 'invalid_request',
+        ];
+    }
+
+    $response->getBody()->write((string) json_encode($responseBody));
+
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return a client error
  */
-$app->match('/clientError', function(Application $app) {
-    return new JsonResponse([
+$app->any('/clientError', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([
         'error' => 'client error',
-    ], 400);
+    ]));
+    return $response
+        ->withHeader('Content-Type', 'appliation/json')
+        ->withStatus(400);
 });
 
 /**
- * Return a server error
+ * @see https://github.com/imbo/behat-api-extension/issues/13
  */
-$app->match('/serverError', function(Application $app) {
-    return new JsonResponse([
-        'error' => 'server error',
-    ], 500);
-});
-
-$app->match('/issue-13', function(Application $app) {
-    return new JsonResponse([
+$app->any('/issue-13', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([
         'customer' => [
             'id' => '12345',
             'name' => 'Behat Testing API',
@@ -163,38 +215,43 @@ $app->match('/issue-13', function(Application $app) {
                 ],
             ],
         ],
-    ]);
+    ]));
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return a response with a custom reason phrase
  */
-$app->get('/customReasonPhrase', function(Application $app) {
-    return (new Response())->setStatusCode(
-        isset($_GET['code']) ? $_GET['code'] : 200,
-        isset($_GET['phrase']) ? $_GET['phrase'] : null
+$app->get('/customReasonPhrase', function(Request $request, Response $response) : Response {
+    $params = $request->getQueryParams();
+
+    return $response->withStatus(
+        !empty($params['code']) ? (int) $params['code'] : 200,
+        !empty($params['phrase']) ? (string) $params['phrase'] : ''
     );
 });
 
 /**
  * Return a response with an empty array
  */
-$app->get('/emptyArray', function(Application $app) {
-    return new JsonResponse([]);
+$app->get('/emptyArray', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode([]));
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return a response with an empty object
  */
-$app->get('/emptyObject', function(Application $app) {
-    return new JsonResponse(new stdClass());
+$app->get('/emptyObject', function(Request $request, Response $response) : Response {
+    $response->getBody()->write((string) json_encode(new stdClass()));
+    return $response->withHeader('Content-Type', 'application/json');
 });
 
 /**
  * Return a response with 403 Forbidden
  */
-$app->get('/403', function(Application $app) {
-    return (new Response())->setStatusCode(403);
+$app->get('/403', function(Request $request, Response $response) : Response {
+    return $response->withStatus(403);
 });
 
 // Run the application
