@@ -1,25 +1,23 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
 
 /**
  * Check if the length of an array is at least a given length
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
 class ArrayMinLength {
     /**
      * Match the min length of an array
      *
      * @param array $array An array
-     * @param int $minLength The expected minimum length of $array
+     * @param int|string $minLength The expected minimum length of $array
      * @throws InvalidArgumentException
-     * @return void
      */
-    public function __invoke($array, $minLength) {
+    public function __invoke($array, $minLength) : bool {
         // Encode / decode to make sure we have a "list"
-        $array = json_decode(json_encode($array));
+        /** @var mixed */
+        $array = json_decode((string) json_encode($array));
 
         if (!is_array($array)) {
             throw new InvalidArgumentException(sprintf(
@@ -38,5 +36,7 @@ class ArrayMinLength {
                 $actualLength
             ));
         }
+
+        return true;
     }
 }
