@@ -1,34 +1,31 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
 
 /**
  * Check if a numeric value is greater than another value
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
 class GreaterThan {
     /**
      * Match a numeric value
      *
-     * @param numeric $number A variable
-     * @param numeric $min The minimum value of $number
+     * @param mixed $number A variable
+     * @param mixed $min The minimum value of $number
      * @throws InvalidArgumentException
-     * @return void
      */
-    public function __invoke($number, $min) {
+    public function __invoke($number, $min) : bool {
         if (!is_numeric($number)) {
             throw new InvalidArgumentException(sprintf(
                 '"%s" is not numeric.',
-                $number
+                (string) $number
             ));
         }
 
         if (!is_numeric($min)) {
             throw new InvalidArgumentException(sprintf(
                 '"%s" is not numeric.',
-                $min
+                (string) $min
             ));
         }
 
@@ -39,5 +36,7 @@ class GreaterThan {
                 $min
             ));
         }
+
+        return true;
     }
 }
