@@ -1,34 +1,31 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
 
 /**
  * Check if a numeric value is less than another value
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
 class LessThan {
     /**
      * Match a numeric value
      *
-     * @param numeric $number A variable
-     * @param numeric $max The max value of $number
+     * @param mixed $number A variable
+     * @param mixed $max The max value of $number
      * @throws InvalidArgumentException
-     * @return void
      */
-    public function __invoke($number, $max) {
+    public function __invoke($number, $max) : bool {
         if (!is_numeric($number)) {
             throw new InvalidArgumentException(sprintf(
                 '"%s" is not numeric.',
-                $number
+                (string) $number
             ));
         }
 
         if (!is_numeric($max)) {
             throw new InvalidArgumentException(sprintf(
                 '"%s" is not numeric.',
-                $max
+                (string) $max
             ));
         }
 
@@ -39,5 +36,7 @@ class LessThan {
                 $max
             ));
         }
+
+        return true;
     }
 }
