@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\Context;
 
 use Behat\Behat\Context\Context;
@@ -6,15 +6,14 @@ use GuzzleHttp\ClientInterface;
 
 /**
  * Api client aware interface
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
 interface ApiClientAwareContext extends Context {
     /**
      * Set the Guzzle client and create a pristine request instance
      *
      * @param ClientInterface $client
+     * @param string $baseUri
      * @return self
      */
-    function setClient(ClientInterface $client);
+    function setClient(ClientInterface $client, string $baseUri);
 }
