@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\Context;
 
 use Imbo\BehatApiExtension\ArrayContainsComparator;
@@ -6,14 +6,11 @@ use Behat\Behat\Context\Context;
 
 /**
  * Array contains comparator aware interface
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
 interface ArrayContainsComparatorAwareContext extends Context {
     /**
      * Set the instance of the array contains comparator
      *
-     * @param ArrayContainsComparator $comparator
      * @return self
      */
     function setArrayContainsComparator(ArrayContainsComparator $comparator);
