@@ -1,11 +1,9 @@
-<?php
+<?php declare(strict_types=1);
 namespace Imbo\BehatApiExtension\Exception;
 
 use Exception;
 
 /**
  * Assertion failure exception
- *
- * @author Christer Edvartsen <cogo@starzinger.net>
  */
 class AssertionFailedException extends Exception {}
